fix(b6): position parameters are continuous, so a DAW can hit an angle exactly
X, Y and Z were AudioParameterInt over [-50, +50] -- 101 steps -- so most
directions simply were not on the grid. Asking for azimuth 30.000 gave x=-25,
y=43 and therefore 30.173517, which is why the B2 null had to be taken against
a Studio reference regenerated at that captured angle rather than at the angle
anyone asked for. They are now AudioParameterFloat over the SAME range.

Float rather than a finer integer scale, deliberately. Hundredths over
[-5000, +5000] would have changed what a stored number MEANS: every session
holding x=43 would load as 0.43 and every object in every existing project
would silently move, which no version hint undoes after the fact. Same range,
finer type, so a saved 43 loads as 43.0 and nothing moves. Host automation is
unaffected either way -- it is recorded normalised 0..1 over an unchanged
range, and the parameter IDs and version hint are untouched, so existing VST3
lanes stay bound and replay to the same positions.

The getters had to widen with them. get{X,Y,Z}Position() returned int and
Panner3DProcessor stored int; left alone, Eclipsa's own monitoring would have
gone on quantising while the KALA export -- which reads the parameters' atomics
directly -- did not, and the two would have drifted apart on every fractional
position. ElevationListener's std::round on the dome write went for the same
reason.

Gate, and it came out better than asked: a scripted REAPER pan at az +30.000
now captures az=30.000000 exactly (asked: within 0.01 deg), and the end-to-end
null against the ORIGINAL, untouched 30.000 deg Studio reference is -inf dBFS
on all twelve channels and SAMPLE-EXACT. No reference regeneration, no residual
LSB. The B2 evidence table gains that row and keeps the old one, because the
old diagnosis is what led here.

+++ common/data_structures/src/AudioElementParameterTree.h
@@ -29,15 +29,19 @@ class AudioElementParameterTree : public juce::AudioProcessorValueTreeState {
             panner, nullptr, AutoParamMetaData::kTreeType,
             AutoParamMetaData::CreateStaticParameterLayout()) {};
 
-  int getXPosition() {
+  // float, not int: the position parameters are continuous (see
+  // ParameterMetaData.h). Returning int here would re-quantise them at every
+  // read and put Eclipsa's own monitoring back out of step with the KALA
+  // export, which reads the same parameters' atomics directly.
+  float getXPosition() {
     return getParameterAsValue(AutoParamMetaData::xPosition).getValue();
   }
 
-  int getYPosition() {
+  float getYPosition() {
     return getParameterAsValue(AutoParamMetaData::yPosition).getValue();
   }
 
-  int getZPosition() {
+  float getZPosition() {
     return getParameterAsValue(AutoParamMetaData::zPosition).getValue();
   }
 
@@ -49,15 +53,15 @@ class AudioElementParameterTree : public juce::AudioProcessorValueTreeState {
     return getParameterAsValue(AutoParamMetaData::unmuteId).getValue();
   }
 
-  void setXPosition(int value) {
+  void setXPosition(float value) {
     getParameterAsValue(AutoParamMetaData::xPosition).setValue(value);
   }
 
-  void setYPosition(int value) {
+  void setYPosition(float value) {
     getParameterAsValue(AutoParamMetaData::yPosition).setValue(value);
   }
 
-  void setZPosition(int value) {
+  void setZPosition(float value) {
     getParameterAsValue(AutoParamMetaData::zPosition).setValue(value);
   }
